feat(Airdrops): airdrop page copy, move comps around
