--- packages/blockchain-wallet-v4-frontend/src/components/Container/index.js
@@ -1,21 +1,0 @@
-import styled from 'styled-components'
-
-const Container = styled.div`
-  width: 100%;
-  margin: 0 auto;
-
-  @media (min-width: 576px) {
-    width: 540px;
-  }
-  @media (min-width: 768px) {
-    width: 720px;
-  }
-  @media (min-width: 992px) {
-    width: 960px;
-  }
-  @media (min-width: 1200px) {
-    width: 1140px;
-  }
-`
-
-export default Container
