--- packages/blockchain-wallet-v4-frontend/src/scenes/BuySell/SfoxCheckout/Content/QuoteInput.js
@@ -1,195 +1,0 @@
-import { path, toUpper } from 'ramda'
-import { Remote } from 'blockchain-wallet-v4/src'
-import FiatConverter from './QuoteInputTemplate'
-import PropTypes from 'prop-types'
-import React, { Component } from 'react'
-
-const WrappedFiatConverter = ({
-  leftVal,
-  leftUnit,
-  rightVal,
-  rightUnit,
-  disabled,
-  onChangeLeft,
-  onChangeRight,
-  limits,
-  reason,
-  quoteR,
-  cryptoMax
-}) => (
-  <FiatConverter
-    value={leftVal}
-    fiat={rightVal}
-    data={Remote.of({
-      unit: leftUnit,
-      currency: rightUnit
-    })}
-    disabled={disabled}
-    handleCoinChange={onChangeLeft}
-    handleFiatChange={onChangeRight}
-    unit={'__required__'}
-    currency={'__required__'}
-    handleBlur={() => {}}
-    handleFocus={() => {}}
-    handleErrorClick={() => {}}
-    meta={{}}
-    limits={limits}
-    quoteR={quoteR}
-    reason={reason}
-    cryptoMax={cryptoMax}
-  />
-)
-
-const convert = {
-  from: {
-    btc: value => Math.floor(parseFloat(value) * 1e8),
-    usd: value => Math.floor(parseFloat(value) * 100)
-  },
-  to: {
-    btc: value => (value / 1e8).toFixed(8).replace(/\.?0+$/, ''),
-    usd: value => String(value).replace(/\.?0+$/, '')
-  }
-}
-
-const otherSide = side => {
-  return side === 'input' ? 'output' : 'input'
-}
-
-class QuoteInput extends Component {
-  constructor (props) {
-    super(props)
-    this.state = {
-      side: 'input',
-      input: this.props.initialAmount,
-      output: '',
-      lastQuoteId: this.props.initialQuoteId,
-      userInput: false
-    }
-
-    this.updateFields = this.updateFields.bind(this)
-  }
-
-  componentDidUpdate (prevProps) {
-    if (prevProps.type !== this.props.type) this.fetchQuoteDebounced()
-    this.props.quoteR.map(quote => {
-      if (quote.id !== path(['quoteR', 'data', 'id'], prevProps)) {
-        this.updateFields(quote)
-      }
-    })
-  }
-
-  updateFields (quote) {
-    if (!this.state.userInput) return null
-    let fiat =
-      this.state.side === 'input' ? this.state.input : quote.quoteAmount
-    let crypto =
-      this.state.side === 'output'
-        ? quote.baseAmount / 1e8
-        : quote.quoteAmount / 1e8
-    this.setState({
-      input: fiat,
-      output: crypto
-    })
-  }
-
-  getQuoteValues = () => {
-    let { spec } = this.props
-    let { side } = this.state
-    return {
-      amt: convert.from[spec[side]](this.state[side]),
-      baseCurrency: toUpper(spec[side]),
-      quoteCurrency: toUpper(spec[otherSide(side)])
-    }
-  }
-
-  handleChangeLeft = event => {
-    this.setState({
-      side: 'input',
-      input: event.target.value,
-      userInput: true
-    })
-    if (!event.target.value) {
-      this.setState({ input: '', output: '' })
-      return null
-    }
-    this.fetchQuoteDebounced()
-  }
-
-  handleChangeRight = event => {
-    this.setState({
-      side: 'output',
-      output: event.target.value,
-      userInput: true
-    })
-    if (!event.target.value) {
-      this.setState({ input: '', output: '' })
-      return null
-    }
-    this.fetchQuoteDebounced()
-  }
-
-  fetchQuoteDebounced = () => {
-    clearTimeout(this.fetchQuoteTimeout)
-    this.fetchQuoteTimeout = setTimeout(this.fetchQuote, this.props.debounce)
-  }
-
-  fetchQuote = () => {
-    let quote = this.getQuoteValues()
-    if (
-      quote.baseCurrency === 'BTC' &&
-      quote.amt / 1e8 > this.props.cryptoMax
-    ) {
-      return null
-    }
-    if (
-      quote.baseCurrency === 'USD' &&
-      quote.amt / 100 > this.props.limits.max
-    ) {
-      return null
-    }
-    this.props.onFetchQuote(quote)
-  }
-
-  render () {
-    let { spec, disabled } = this.props
-    let { input, output, userInput } = this.state
-
-    return (
-      <WrappedFiatConverter
-        leftVal={input}
-        leftUnit={toUpper(spec.input)}
-        onChangeLeft={this.handleChangeLeft}
-        rightVal={output}
-        rightUnit={toUpper(spec.output)}
-        onChangeRight={this.handleChangeRight}
-        limits={this.props.limits}
-        disabled={disabled}
-        userInput={userInput}
-        quoteR={this.props.quoteR}
-        reason={this.props.reason}
-        cryptoMax={this.props.cryptoMax}
-      />
-    )
-  }
-}
-
-QuoteInput.propTypes = {
-  quoteR: PropTypes.any.isRequired,
-  initialAmount: PropTypes.string,
-  debounce: PropTypes.number,
-  spec: PropTypes.shape({
-    method: PropTypes.string,
-    output: PropTypes.string,
-    input: PropTypes.string
-  }).isRequired,
-  onFetchQuote: PropTypes.func.isRequired,
-  initialQuoteId: PropTypes.string
-}
-
-QuoteInput.defaultProps = {
-  initialAmount: '',
-  debounce: 500,
-  initialQuoteId: null
-}
-
-export default QuoteInput
